backend/requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9.9,<3.0
redis>=5.0.0
django-redis>=5.4.0
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
pytest-django>=4.8.0
pytest-asyncio>=0.23.0
# -- Nuevas dependencias para Autenticación tipo Supabase --
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.0
# -- Tareas programadas (Limpieza de tokens) --
celery>=5.3.6
# -- Soporte para Imágenes (Avatares) --
Pillow>=10.2.0
